requirements.txt
requests
quart
hypercorn
pandas
pyyaml
plotly